cmd/edit: move editText to the common file
editText() is being used not only for issues, cmd/edit_common.go is a
better place for it. Its commit description was also misleading.
